requirements.txt
flask 
gunicorn
requests
google-generativeai
python-dotenv